Improve network setup with bridging and ntpdate
The uap0 and eth1 interfaces are bridged as br0, to enable wired and
wireless devices to communicate seamlessly.

As soon as the network is up on eth0 (via DHCP), ntpdate is used to
set the time/date of the machine in a more precise way than just
using the last saved time.